requirements.txt
# Core dependencies
aiohttp==3.9.1
aiofiles==23.2.1
asyncio-mqtt==0.16.1
uvloop==0.19.0

# Web interface
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
websockets==12.0

# Database and storage
sqlite3-to-postgres==0.4.4
aiosqlite==0.19.0
redis==5.0.1

# LLM APIs
openai==1.3.7
google-generativeai==0.3.2
anthropic==0.7.7
requests==2.31.0

# Git and version control
GitPython==3.1.40
pygit2==1.13.3

# Code analysis
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-c==0.20.6
tree-sitter-cpp==0.20.3
ast-decompiler==0.7.0
astunparse==1.6.3

# Testing and validation
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2

# MCP and communication
pydantic==2.5.0
msgpack==1.0.7
websocket-client==1.6.4

# Utilities
click==8.1.7
rich==13.7.0
tabulate==0.9.0
psutil==5.9.6
pyyaml==6.0.1
toml==0.10.2
configparser==6.0.0

# Security and authentication
cryptography==41.0.8
passlib==1.7.4
python-jose==3.3.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# ARM/aarch64 optimized packages
lxml==4.9.3
numpy==1.24.4

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0